cat-con/requirements.txt
streamlit
pymongo
pandas
pillow
rapidfuzz
python-dotenv
dnspython